moved RRE to src directory
